refactor: gather the legacy sled types under sled_compat
# Summary

`key_spaces`, `state` and the two metadata modules under `state_machine` merge
into one `sled_compat` module. `MetaSnapshotId` moves to `sm_v003`, and the
`state_machine` module name is now free.

# Details

Three modules described the same thing from different angles: the pre-V004
sled layout. Nothing writes it any more; it exists so the upgrader can read
data left by an older version and so the exporter can keep emitting its entry
format. Scattered across the crate that was invisible, and it left the name
`state_machine` on a module holding only legacy metadata types while the
actual state machine lived under `sm_v003`.

Grouping them names the boundary and makes the set deletable in one piece when
V003 compatibility is dropped.

`MetaSnapshotId` was the odd member of `state_machine`: it identifies a V004
snapshot and is used only by the snapshot writer, loader and store. It moves
next to them.

Public paths change accordingly:

    key_spaces::*             -> sled_compat::key_spaces::*
    state::*                  -> sled_compat::*
    state_machine::MetaSnapshotId -> sm_v003::MetaSnapshotId
