+++ crates/server/raft-store/src/lib.rs
@@ -48,14 +48,12 @@ pub mod data_version;
 pub mod db_exporter;
 pub mod header;
 pub mod immutable_compactor;
-pub mod key_spaces;
 pub mod leveled_store;
 pub mod ondisk;
 pub mod raft_log_v004;
+pub mod sled_compat;
 pub mod sm_v003;
 pub mod snapshot_config;
-pub mod state;
-pub mod state_machine;
 pub mod state_machine_api_ext;
 pub(crate) mod testing;
 pub mod utils;
