+++ crates/server/raft-store/src/sm_v003/snapshot_id.rs
@@ -155,7 +155,7 @@ mod tests {
 
     use databend_meta_types::raft_types::new_log_id;
 
-    use crate::state_machine::snapshot_id::MetaSnapshotId;
+    use crate::sm_v003::MetaSnapshotId;
 
     #[test]
     fn test_meta_snapshot_id() -> anyhow::Result<()> {
