+++ crates/server/raft-store/src/sm_v003/snapshot_loader.rs
@@ -22,10 +22,10 @@ use log::info;
 use log::warn;
 use openraft::SnapshotId;
 
+use crate::sm_v003::MetaSnapshotId;
 use crate::sm_v003::SnapshotStoreError;
 use crate::sm_v003::open_snapshot::OpenSnapshot;
 use crate::snapshot_config::SnapshotConfig;
-use crate::state_machine::MetaSnapshotId;
 
 /// Load snapshot from persisted storage.
 ///
