+++ crates/server/service/src/store/store.rs
@@ -24,15 +24,15 @@ use databend_meta_raft_store::MetaStartupError;
 use databend_meta_raft_store::config::RaftConfig;
 use databend_meta_raft_store::db_exporter::DBExporter;
 use databend_meta_raft_store::header::Header;
-use databend_meta_raft_store::key_spaces::RaftStoreEntry;
 use databend_meta_raft_store::ondisk::TREE_HEADER;
 use databend_meta_raft_store::raft_log_v004;
 use databend_meta_raft_store::raft_log_v004::Cw;
 use databend_meta_raft_store::raft_log_v004::RaftLogV004;
 use databend_meta_raft_store::raft_log_v004::util;
+use databend_meta_raft_store::sled_compat::key_spaces::RaftStoreEntry;
+use databend_meta_raft_store::sm_v003::MetaSnapshotId;
 use databend_meta_raft_store::sm_v003::SMV003;
 use databend_meta_raft_store::sm_v003::SnapshotStoreV004;
-use databend_meta_raft_store::state_machine::MetaSnapshotId;
 use databend_meta_runtime_api::SpawnApi;
 use databend_meta_snapshot_db::DB;
 use databend_meta_snapshot_db::DBStat;
